backend/requirements.txt
pydantic
fastapi
python-dotenv
uvicorn
tiktoken
boto3
openai
mistralai
docling
asyncio